BUG/CoW: is_range_indexer can't handle very large arrays (#53672)
* BUG: is_range_indexer can't handle very large arrays

* fix test on 32-bit

+++ pandas/_libs/lib.pyx
@@ -668,7 +668,7 @@ ctypedef fused int6432_t:
 
 @cython.wraparound(False)
 @cython.boundscheck(False)
-def is_range_indexer(ndarray[int6432_t, ndim=1] left, int n) -> bool:
+def is_range_indexer(ndarray[int6432_t, ndim=1] left, Py_ssize_t n) -> bool:
     """
     Perform an element by element comparison on 1-d integer arrays, meant for indexer
     comparisons
